+++ packages/react-composites/src/composites/CallComposite/components/buttons/Camera.tsx
@@ -1,0 +1,27 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { CameraButton, ControlBarButtonStyles } from '@internal/react-components';
+import React, { useMemo } from 'react';
+import { usePropsFor } from '../../hooks/usePropsFor';
+import { CallControlDisplayType } from '../../types/CallControlOptions';
+import { concatButtonBaseStyles } from './styles';
+
+export const Camera = (props: {
+  displayType?: CallControlDisplayType;
+  styles?: ControlBarButtonStyles;
+  splitButtonsForDeviceSelection?: boolean;
+}): JSX.Element => {
+  const cameraButtonProps = usePropsFor(CameraButton);
+  const styles = useMemo(() => concatButtonBaseStyles(props.styles ?? {}), [props.styles]);
+  return (
+    <CameraButton
+      data-ui-id="call-composite-camera-button"
+      {...cameraButtonProps}
+      showLabel={props.displayType !== 'compact'}
+      styles={styles}
+      /* @conditional-compile-remove-from(stable) meeting-composite control-bar-split-buttons */
+      enableDeviceSelectionMenu={props.splitButtonsForDeviceSelection}
+    />
+  );
+};
